AVO-3537: set default login tab depending on secured route for assignment details and quick lane detail views
